cron.php should not be runnable by users

+++ cron.php
@@ -1,20 +1,27 @@
 #!/usr/bin/env php
 <?php
 
-$root = dirname(__FILE__);
-require_once ($root . '/classes/db/Database.php');
+if (php_sapi_name() === 'cli') {
 
-$database = new Database();
-$conn = $database->conn;
+	$root = dirname(__FILE__);
+	require_once ($root . '/classes/db/Database.php');
 
-// delete expired verify entries
-$stmt = $conn->prepare('DELETE FROM verify WHERE ExpiresAt < NOW()');
-$stmt->execute();
-echo $stmt->rowCount() . " verifications expired\n";
+	$database = new Database();
+	$conn = $database->conn;
 
-// delete expired sessions
-$stmt = $conn->prepare('DELETE FROM sessions WHERE ExpiresAt < NOW()');
-$stmt->execute();
-echo $stmt->rowCount() . " sessions expired\n";
+	// delete expired verify entries
+	$stmt = $conn->prepare('DELETE FROM verify WHERE ExpiresAt < NOW()');
+	$stmt->execute();
+	echo $stmt->rowCount() . " verifications expired\n";
+
+	// delete expired sessions
+	$stmt = $conn->prepare('DELETE FROM sessions WHERE ExpiresAt < NOW()');
+	$stmt->execute();
+	echo $stmt->rowCount() . " sessions expired\n";
+
+} else {
+	header("Content-Type: text/plain");
+	echo "\ncron.php can only be run from CLI\n";
+}
 
 ?>
